s/processd: rewrite the remaining code, now we just need to cleanup and implement the rest of the new features
